+++ phpBB2/ctracker/admin/acp_module_settings.php
@@ -45,7 +45,8 @@ if( isset($HTTP_POST_VARS['submit']) )
 		'detect_misconfiguration' => array(0, 1), 'spam_attack_boost' => array(0, 1),
 		'spam_keyword_det' => array(0, 2), 'request_limit_enabled' => array(0, 1),
 		'request_limit_login' => array(5, 100), 'request_limit_register' => array(1, 50),
-		'request_limit_write' => array(20, 500), 'request_limit_upload' => array(1, 100)
+		'request_limit_account' => array(1, 100), 'request_limit_write' => array(20, 500),
+		'request_limit_upload' => array(1, 100)
 	);
 	foreach ($setting_ranges as $setting_name => $range)
 	{
@@ -105,6 +106,7 @@ $configuration['spam_keyword_det'] 		  = $adminclass->ct_keyword_b_block($ctrack
 $configuration['request_limit_enabled']  = $adminclass->ct_generate_on_off($ctracker_config->settings['request_limit_enabled']);
 $configuration['request_limit_login']    = $adminclass->ct_generate_number_field(5, 100, $ctracker_config->settings['request_limit_login']);
 $configuration['request_limit_register'] = $adminclass->ct_generate_number_field(1, 50, $ctracker_config->settings['request_limit_register']);
+$configuration['request_limit_account']  = $adminclass->ct_generate_number_field(1, 100, $ctracker_config->settings['request_limit_account']);
 $configuration['request_limit_write']    = $adminclass->ct_generate_number_field(20, 500, $ctracker_config->settings['request_limit_write']);
 $configuration['request_limit_upload']   = $adminclass->ct_generate_number_field(1, 100, $ctracker_config->settings['request_limit_upload']);
 
@@ -217,6 +219,8 @@ $template->assign_vars(array(
 		'L_EXP_40'	   => $lang['ctracker_settings_e40'],
 		'L_MOD_41'     => $lang['ctracker_settings_m41'],
 		'L_EXP_41'	   => $lang['ctracker_settings_e41'],
+		'L_MOD_42'     => $lang['ctracker_settings_m42'],
+		'L_EXP_42'	   => $lang['ctracker_settings_e42'],
 		
 		'CAT_ICON_1'   => $phpbb_root_path . $images['ctracker_icon_set_1'],
 		'CAT_ICON_2'   => $phpbb_root_path . $images['ctracker_icon_set_2'],
@@ -267,6 +271,7 @@ $template->assign_vars(array(
 		'S_OUTPUT_39'  => $configuration['request_limit_register'],
 		'S_OUTPUT_40'  => $configuration['request_limit_write'],
 		'S_OUTPUT_41'  => $configuration['request_limit_upload'],
+		'S_OUTPUT_42'  => $configuration['request_limit_account'],
 		
 		'S_FORM_TOKEN' => phpbb_admin_session_field(),
 		'S_FORM_ACTION' => append_sid('admin_cracker_tracker.' . $phpEx . '?modu=9'))