+++ phpBB2/ctracker/classes/class_ct_database.php
@@ -76,6 +76,7 @@ class ct_database
 			'request_limit_enabled' => '1',
 			'request_limit_login' => '30',
 			'request_limit_register' => '10',
+			'request_limit_account' => '20',
 			'request_limit_write' => '120',
 			'request_limit_upload' => '30'
 		);